refactor(gui): separate base window classes from session-related gui
