+++ .gitignore
@@ -1,7 +1,6 @@
 __pycache__/
 test_dataset/
 notes/
-testing/
 .venv/
 .python-version
